Rename some class name
