+++ shared/src/commonMain/kotlin/com/blucky8649/eungabi/SampleApp.kt
@@ -35,9 +35,9 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.unit.dp
-import com.blucky8649.eungabi.navigation.experimental.EunGabiController
-import com.blucky8649.eungabi.navigation.experimental.EunGabiNavHost
-import com.blucky8649.eungabi.navigation.experimental.rememberEunGabiController
+import com.blucky8649.eungabi.navigation.EunGabiController
+import com.blucky8649.eungabi.navigation.EunGabiNavHost
+import com.blucky8649.eungabi.navigation.rememberEunGabiController
 import com.blucky8649.sample.resources.Res
 import com.blucky8649.sample.resources.ic_dy
 import org.jetbrains.compose.resources.painterResource
@@ -130,7 +130,7 @@ fun BackStackTracker(navController: EunGabiController) {
     Box(Modifier.fillMaxSize()) {
         LazyRow(Modifier.fillMaxWidth().align(Alignment.BottomCenter)) {
             items(backStack.size) {
-                val route = backStack[it].destination.route
+                val route = backStack[it].eunGabiDestination.route
                 Text(route, modifier = Modifier.background(Color.LightGray))
                 Spacer(Modifier.width(10.dp))
             }
